tests/test_clingen_client.py
"""Tests for ClinGen client gene extraction."""
import pytest
from app.pipeline.sources.clingen_client import extract_gene_from_query


def test_extract_brca1_from_natural_query():
    """Should find BRCA1 in a natural clinical question."""
    result = extract_gene_from_query("What is the clinical significance of rs879254116 in BRCA1?")
    assert result == "BRCA1"


def test_extract_gene_returns_none_for_no_gene():
    """Query with no gene symbol should return None."""
    result = extract_gene_from_query("What is the weather like today?")
    assert result is None


def test_extract_gene_ignores_acmg_acronym():
    """ACMG is a common acronym, not a gene — should not be returned."""
    result = extract_gene_from_query("What are the ACMG criteria for variant classification?")
    assert result != "ACMG"


def test_extract_gene_ignores_nccn_acronym():
    """NCCN is a common acronym, not a gene — should not be returned."""
    result = extract_gene_from_query("What does the NCCN guideline recommend?")
    assert result != "NCCN"


def test_extract_gene_ignores_vus_acronym():
    """VUS is a classification term, not a gene."""
    result = extract_gene_from_query("Is this variant a VUS?")
    assert result != "VUS"


def test_extract_multiple_genes_returns_first():
    """When multiple gene symbols appear, should return the first one."""
    result = extract_gene_from_query("Compare BRCA1 and TP53 pathogenic variants")
    assert result in ("BRCA1", "TP53")  # either is acceptable as "first found"


def test_extract_unusual_gene_not_in_old_allowlist():
    """CDH1 was NOT in the old 28-gene allowlist — regex should find it (H-3 fix)."""
    result = extract_gene_from_query("Is there a CDH1 pathogenic variant?")
    assert result == "CDH1"


def test_extract_gene_handles_lowercase_query():
    """Gene symbols in lowercase context should still be found (they're uppercase tokens)."""
    result = extract_gene_from_query("check brca2 variant for pathogenicity")
    # BRCA2 appears as 'brca2' in lowercase — the regex looks for uppercase tokens
    # This may return None depending on implementation, which is acceptable
    # The important thing is it doesn't crash
    assert result is None or result == "BRCA2"
