requirements.txt
# ── Web Framework ─────────────────────────────────────────────────
fastapi==0.131.0
uvicorn==0.41.0

# ── Data Validation ───────────────────────────────────────────────
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.13.1

# ── Async ─────────────────────────────────────────────────────────
starlette==0.52.1

# ── Database (PostgreSQL + pgvector) ──────────────────────────────
psycopg2-binary==2.9.11
sqlalchemy==2.0.46
pgvector==0.3.6

# ── AI / LLM ──────────────────────────────────────────────────────
langchain==1.2.10
langchain-community==0.4.1
langchain-postgres==0.0.17
langgraph==1.0.10
sentence-transformers==5.2.3
ollama==0.6.1

# ── HTTP Client (for ClinGen API) ─────────────────────────────────
httpx==0.28.1

# ── PDF Processing ────────────────────────────────────────────────
pymupdf==1.27.1
pymupdf4llm==0.3.4
docling==2.77.0